Finished fixing issues with wide char to UTF-8 conversions, also fixed most norm compliance issues. Still need to go over macro use and formatting floating point conversions

+++ srcs/parseunum.c
@@ -58,7 +58,7 @@ void		parseunum(t_format *flags, va_list *args)
 	}
 	else if (LEN == 2)
 	{
-		flags->arg = (unsigned long long*)ft_memalloc(sizeof(unsigned long long));
+		flags->arg = (unsigned long*)ft_memalloc(sizeof(unsigned long long));
 		*(unsigned long long*)(flags->arg) = va_arg(*args, unsigned long long);
 		flags->argsize = sizeof(unsigned long long);
 	}
